requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
